requirements.txt
asgiref
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
python-dotenv
sqlparse
whitenoise
